requirements.txt
rich
nodriver
python-dotenv
requests
wasmtime